Control.Comonad.Cofree2

Documentation

exampleUses :: ()

Example uses:

 lines
 comments
 strings
 matched: matched stack
 spaces
 alpha
 numeric
 symbols
 build special symbols: ',` etc, into applications

 add type for unresolved precidence problem including resoltions to applications

 split into: var, app
 split into: , incl. derived general resolutions (e.g. (.. (f x)) => f :: a -> b, x :: a)

scanlBicofree :: (c -> a -> c) -> c -> Bicofree Maybe a b -> Maybe (Cofree Maybe (Env c b))

Scan left, recording results in Env

pushScanlCofree :: (c -> a -> c) -> c -> Cofree Maybe (Either a b) -> Maybe (Cofree Maybe (Env c (Cofree Maybe b)))

Generate a Bicofree from eitherCofreeMaybe2, then scanlBicofree with foldl over the internal Cofree

linesCofree :: Enum c => Cofree Maybe (Either a b) -> Maybe (Cofree Maybe (Env c (Cofree Maybe b)))

Lines c a = Cofree Maybe (Env c a) Cofree Maybe (Either a b) -> Maybe (Lines c (Cofree Maybe b))

eitherCofreeMaybeP :: (a -> Bool) -> Cofree Maybe a -> Cofree Maybe (Either a a)

Lefts are matches

eitherCofreeLines :: Comonad f => Cofree Maybe (f Char) -> Cofree Maybe (Either (f Char) (f Char))

Lefts are newlines

eitherCofreeSpaces :: Comonad f => Cofree Maybe (f Char) -> Cofree Maybe (Either (f Char) (f Char))

Collects spaces

eitherCofreeMaybeBegin2 :: (a -> a -> Bool) -> Cofree Maybe a -> Cofree Maybe (Either a a)

Begins when predicate matched, never ends

eitherCofreeComment :: Comonad f => Cofree Maybe (f Char) -> Cofree Maybe (Either (f Char) (f Char))

Lefts are comments (ending the line)

eitherCofreeMaybeBeginEnd :: (a -> Bool) -> (a -> Bool) -> Cofree Maybe a -> Cofree Maybe (Either a a)

begin starts the Left, end ends it

eitherCofreeMaybeBeginEnd2 :: (a -> Bool) -> (a -> a -> Bool) -> Cofree Maybe a -> Cofree Maybe (Either a a)

begin starts the Left, end ends it

eitherCofreeString :: Comonad f => Cofree Maybe (f Char) -> Cofree Maybe (Either (f Char) (f Char))

Parses a string

eitherCofreeMaybe2 :: Cofree Maybe (Either a b) -> Bicofree Maybe (Cofree Maybe a) (Cofree Maybe b)

Use Bicofree to push through a representation of a non-empty list

fromBicofree :: Functor f => Bicofree f a a -> Cofree f a

See fromCofree2

newtype Bicofree f a b

Cofree2, when we don't know whether a or b occurs first

Constructors

Bicofree

Fields

runBicofree :: Either (Cofree2 f b a) (Cofree2 f a b)

fromCofree2 :: Functor f => Cofree2 f a a -> Cofree f a

Rewriting (:<<) -> (:<) gives Cofree

unwrap2 :: Functor f => Cofree2 f a b -> f (f (Cofree2 f a b))

Unwrap two layers of a Cofree2 by dropping their values

data Cofree2 f a b where

Constructors

(:<<) :: b -> f (Cofree2 f b a) -> Cofree2 f a b infixr 4